Survey moved out from internal story
