+++ packages/storybook/stories/Survey/components/SurveyExample.tsx
@@ -5,7 +5,7 @@ import { CallSurvey } from '@azure/communication-calling';
 import { CallSurveyImprovementSuggestions } from '@azure/communication-react';
 import { Text, PrimaryButton, Stack, Theme, mergeStyles, useTheme } from '@fluentui/react';
 import React, { useState } from 'react';
-import { SurveyContent } from '../../../../../react-composites/src/composites/common/SurveyContent';
+import { SurveyContent } from '../../../../react-composites/src/composites/common/SurveyContent';
 
 export const SurveyExample = (): JSX.Element => {
   const [ratings, setRatings] = useState(0);
